rename the docs version too
